refactor: Extract widgets
