+++ lib/src/story_page_wrapper.dart
@@ -1,8 +1,8 @@
-import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:storybook_flutter/src/breakpoint.dart';
-import 'package:storybook_flutter/src/story.dart';
+import 'package:storybook_flutter/src/contents.dart';
+import 'package:storybook_flutter/src/current_story.dart';
 import 'package:storybook_flutter/src/story_provider.dart';
 
 class StoryPageWrapper extends StatelessWidget {
@@ -18,16 +18,16 @@ class StoryPageWrapper extends StatelessWidget {
     final isFullPage = data.isFullPage;
 
     if (isFullPage) {
-      return Scaffold(body: _buildStory(context, story));
+      return const Scaffold(body: CurrentStory());
     }
 
     return Scaffold(
       drawer: _shouldDisplayDrawer(context)
-          ? const Drawer(child: _Contents())
+          ? const Drawer(child: Contents())
           : null,
       appBar: AppBar(title: Text(story?.name ?? 'Storybook')),
       body: _shouldDisplayDrawer(context)
-          ? _buildStory(context, story)
+          ? const CurrentStory()
           : Row(
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: <Widget>[
@@ -41,56 +41,11 @@ class StoryPageWrapper extends StatelessWidget {
                     color: Theme.of(context).cardColor,
                   ),
                   width: 200,
-                  child: const _Contents(),
+                  child: const Contents(),
                 ),
-                Expanded(child: _buildStory(context, story)),
+                const Expanded(child: CurrentStory()),
               ],
             ),
     );
   }
-
-  Widget _buildStory(BuildContext context, Story? story) =>
-      story ?? const Center(child: Text('Select story'));
-
-  void _onStoryTap(BuildContext context, Story story) {
-    if (_shouldDisplayDrawer(context)) Navigator.of(context).pop();
-    Navigator.pushReplacementNamed(context, '/stories/${story.path}');
-  }
-}
-
-class _Contents extends StatefulWidget {
-  const _Contents({Key? key}) : super(key: key);
-
-  @override
-  _ContentsState createState() => _ContentsState();
-}
-
-class _ContentsState extends State<_Contents> {
-  @override
-  Widget build(BuildContext context) {
-    final grouped = context.watch<List<Story>>().groupListsBy((s) => s.section);
-    final sections = grouped.keys
-        .where((k) => k.isNotEmpty)
-        .map((k) => _buildSection(k, grouped[k]!));
-    final stories = (grouped[''] ?? []).map(_buildStoryTile);
-    return ListTileTheme(
-      style: ListTileStyle.drawer,
-      child: ListView(children: [...sections, ...stories]),
-    );
-  }
-
-  Widget _buildStoryTile(Story story) => ListTile(
-        title: Text(story.name),
-        onTap: () {
-          Navigator.pushReplacementNamed(context, '/stories/${story.path}');
-        },
-        selected: story == context.watch<StoryProvider>().currentStory,
-      );
-
-  Widget _buildSection(String title, Iterable<Story> stories) => ExpansionTile(
-        title: Text(title),
-        initiallyExpanded:
-            stories.contains(context.watch<StoryProvider>().currentStory),
-        children: stories.map(_buildStoryTile).toList(),
-      );
 }
